+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.5.5
 * Changed kill payouts to be based on creature spawn type, not distance from spawners
 * Fixed duplicate honorific on /jobs reload
+* Improved database performance
 
 ## 2.5.4
 * Yet another crafting NPE fix